Change script to have options to add/remove allowed origins
